+++ jigoshop.php
@@ -18,11 +18,11 @@
  * Plugin Name:         Jigoshop eCommerce
  * Plugin URI:          http://www.jigoshop.com/
  * Description:         Jigoshop, a WordPress eCommerce plugin that works.
- * Author:              Jigoshop
+ * Author:              Jigo Ltd
  * Author URI:          http://www.jigoshop.com
- * Version:             2.1.5
+ * Version:             2.1.6
  * Requires at least:   4.0
- * Tested up to:        4.8.1
+ * Tested up to:        4.8.2
  * Text Domain:         jigoshop
  * Domain Path:         /languages/
  *